Move eventHandlex index to own folder outside utils
