+++ ui/ui-components/eventHandlers/index.ts
@@ -1,0 +1,35 @@
+import { ErrorHandler } from "../utils/errorHandler";
+import { SuccessHandler } from "../utils/successHandler";
+import { AppHandler } from "../utils/appHandler";
+import { DestinationHandler } from "../utils/destinationHandler";
+import { GroupHandler } from "../utils/groupHandler";
+import { RecordHandler } from "../utils/recordHandler";
+import { RecordsHandler } from "../utils/recordsHandler";
+import { PropertiesHandler } from "../utils/propertiesHandler";
+import { RunsHandler } from "../utils/runsHandler";
+import { ScheduleHandler } from "../utils/scheduleHandler";
+import { SessionHandler } from "../utils/sessionHandler";
+import { SetupStepHandler } from "../utils/setupStepsHandler";
+import { StatusHandler } from "../utils/statusHandler";
+import { SourceHandler } from "../utils/sourceHandler";
+import { TeamHandler } from "../utils/teamHandler";
+import { TeamMemberHandler } from "../utils/teamMembersHandler";
+import { UploadHandler } from "../utils/uploadHandler";
+
+export const successHandler = new SuccessHandler();
+export const errorHandler = new ErrorHandler();
+export const appHandler = new AppHandler();
+export const destinationHandler = new DestinationHandler();
+export const groupHandler = new GroupHandler();
+export const recordHandler = new RecordHandler();
+export const recordsHandler = new RecordsHandler();
+export const propertiesHandler = new PropertiesHandler();
+export const runsHandler = new RunsHandler();
+export const scheduleHandler = new ScheduleHandler();
+export const sessionHandler = new SessionHandler();
+export const setupStepHandler = new SetupStepHandler();
+export const statusHandler = new StatusHandler();
+export const sourceHandler = new SourceHandler();
+export const teamHandler = new TeamHandler();
+export const teamMemberHandler = new TeamMemberHandler();
+export const uploadHandler = new UploadHandler();
