--- ui/ui-components/utils/eventHandlers.ts
@@ -1,35 +1,0 @@
-import { ErrorHandler } from "./errorHandler";
-import { SuccessHandler } from "./successHandler";
-import { AppHandler } from "./appHandler";
-import { DestinationHandler } from "./destinationHandler";
-import { GroupHandler } from "./groupHandler";
-import { RecordHandler } from "./recordHandler";
-import { RecordsHandler } from "./recordsHandler";
-import { PropertiesHandler } from "./propertiesHandler";
-import { RunsHandler } from "./runsHandler";
-import { ScheduleHandler } from "./scheduleHandler";
-import { SessionHandler } from "./sessionHandler";
-import { SetupStepHandler } from "./setupStepsHandler";
-import { StatusHandler } from "./statusHandler";
-import { SourceHandler } from "./sourceHandler";
-import { TeamHandler } from "./teamHandler";
-import { TeamMemberHandler } from "./teamMembersHandler";
-import { UploadHandler } from "./uploadHandler";
-
-export const successHandler = new SuccessHandler();
-export const errorHandler = new ErrorHandler();
-export const appHandler = new AppHandler();
-export const destinationHandler = new DestinationHandler();
-export const groupHandler = new GroupHandler();
-export const recordHandler = new RecordHandler();
-export const recordsHandler = new RecordsHandler();
-export const propertiesHandler = new PropertiesHandler();
-export const runsHandler = new RunsHandler();
-export const scheduleHandler = new ScheduleHandler();
-export const sessionHandler = new SessionHandler();
-export const setupStepHandler = new SetupStepHandler();
-export const statusHandler = new StatusHandler();
-export const sourceHandler = new SourceHandler();
-export const teamHandler = new TeamHandler();
-export const teamMemberHandler = new TeamMemberHandler();
-export const uploadHandler = new UploadHandler();
